simplify layouts
