+++ src/app/[locale]/auth/layout.tsx
@@ -1,35 +1,7 @@
-import { Roboto } from 'next/font/google'
-import { NextIntlClientProvider } from 'next-intl'
-import { getMessages } from 'next-intl/server'
-import '../../globals.css'
-
-export const metadata = {
-	title: 'Next.js',
-	description: 'Generated by Next.js',
-}
-
-export const roboto = Roboto({
-	subsets: ['latin'],
-	display: 'swap',
-	weight: ['100', '300', '400', '400', '500', '700', '900'],
-})
-
 export default async function RootLayout({
 	children,
-	params: { locale },
 }: Readonly<{
 	children: React.ReactNode
-	params: { locale: string }
 }>) {
-	const messages = await getMessages()
-
-	return (
-		<html lang={locale}>
-			<body className={`${roboto.className} bg-brand-100`}>
-				<NextIntlClientProvider messages={messages}>
-					<main className='grid-layout h-screen items-center'>{children}</main>
-				</NextIntlClientProvider>
-			</body>
-		</html>
-	)
+	return <main className='grid-layout h-screen items-center'>{children}</main>
 }
